+++ include/impl/matrix_io.tpp
@@ -8,9 +8,9 @@
 template <typename T>
 std::ostream &operator<<(std::ostream &os, const matrix<T> &m)
 {
-    for (int i = 0; i < m.rows(); ++i)
+    for (size_t i = 0; i < m.rows(); ++i)
     {
-        for (int j = 0; j < m.cols(); ++j)
+        for (size_t j = 0; j < m.cols(); ++j)
         {
             os << m(i, j) << " ";
         }
@@ -21,12 +21,12 @@ std::ostream &operator<<(std::ostream &os, const matrix<T> &m)
 template <typename T>
 std::istream &operator>>(std::istream &is, matrix<T> &m)
 {
-    int rows, cols;
+    size_t rows, cols;
     is >> rows >> cols;
     m = matrix<T>(rows, cols);
-    for (int i = 0; i < m.rows(); ++i)
+    for (size_t i = 0; i < m.rows(); ++i)
     {
-        for (int j = 0; j < m.cols(); ++j)
+        for (size_t j = 0; j < m.cols(); ++j)
         {
             is >> m(i, j);
         }
